add error handling in controllers and middlewares

See merge request college/developer4/ava-land/404/backend!8

+++ apps/music/controllers/music.controller.js
@@ -7,6 +7,7 @@ const sendResponse = require('../../../modules/handler/response.handler');
 const MusicRepository = require("../repositories/music.repository");
 const musicRepository = new MusicRepository();
 const { date } = require('../../../modules/utility');
+const Music = require('../models/music.model');
 
 
 class MusicController {
@@ -17,7 +18,6 @@ class MusicController {
             const musicJson = {
                 "music-id": music.MusicID,
                 "album-name": music.AlbumName,
-                "music-id": music.MusicID,
                 "name": music.MusicName,
                 "title": music.MusicTitle,
                 "poster": music.MusicPoster,
@@ -49,21 +49,23 @@ class MusicController {
                 sendResponse(res, statusCode.OK, { "Content-Type": "application/json" }, JSON.stringify(this.#print(musics)));
             }
         } catch (error) {
-            logger.error('getMusics: ' + error);
+            logger.error(`${req.url}: ${error}`);
             throw error;
         }
     };
 
     root = async (req, res) => {
-        sendResponse(res, statusCode.OK, { 'Content-Type': 'text/html' },
-            `<html><body>
-            <h2>With Node.js <code>"http"</code> module</h2>
-            <form action="/music/upload" enctype="multipart/form-data" method="post">
-            <div>Text field title: <input type="text" name="title" /></div>
-            <div>File: <input type="file" name="multipleFiles" multiple="multiple" /></div>
-            <input type="submit" value="Upload" />
-            </form>
-            </body></html>`);
+        const page = `<html><body>
+        <h2>With Node.js <code>"http"</code> module</h2>
+        <form action="/music/upload" enctype="multipart/form-data" method="post">
+        <div>Text field title: <input type="text" name="title" /></div>
+        <div>File: <input type="file" name="multipleFiles" multiple="multiple" /></div>
+        <input type="submit" value="Upload" />
+        </form>
+        </body></html>`;
+        res.writeHead(statusCode.OK, { 'Content-Type': 'text/html' });
+        return res.end(page);
+        // return sendResponse(res, statusCode.OK, { 'Content-Type': 'text/html' }, page);
     };
 
     upload = async (req, res) => {
@@ -82,7 +84,7 @@ class MusicController {
                 req.body = files;
 
                 if (error) {
-                    logger.error(error);
+                    logger.error(`formidable: ${error}`);
                     return sendResponse(res, error.httpCode || 400, { 'Content-Type': 'text/plain' }, error);
                 }
                 // sendResponse(res, statusCode.OK, { 'Content-Type': 'application/json' }, JSON.stringify({ fields, files }));
@@ -91,7 +93,7 @@ class MusicController {
 
             return;
         } catch (error) {
-            logger.error(error);
+            logger.error(`${req.url}: ${error}`);
             throw error;
         }
     };
@@ -104,17 +106,14 @@ class MusicController {
             if (!newFilename)
                 return sendResponse(res, statusCode.BAD_REQUEST, { "Content-Type": "application/json" }, 'Invalid parameters!');
      
-            const newMusic = {
-                MusicURL: newFilename
-                // MusicMimeType: mimetype
-            };
+            const newMusic = new Music(0, '', null, '', '', '', newFilename);
             const music = await musicRepository.add(newMusic, req.UserID);
             if (!music)
                 return sendResponse(res, statusCode.NOT_FOUND, { "Content-Type": "application/json" }, 'Could Not Create');
             else
                 return sendResponse(res, statusCode.OK, { 'Content-Type': 'application/json' }, JSON.stringify(this.#print([music])));
         } catch (error) {
-            logger.error('createMusic: ' + error);
+            logger.error(`${req.url}: ${error}`);
             throw error;
         }
     };
@@ -123,18 +122,20 @@ class MusicController {
         try {
             const { id } = req.querystring;
             const { body } = req;
+            if (!id || !body)
+                return sendResponse(res, statusCode.BAD_REQUEST, { "Content-Type": "application/json" }, 'Invalid parameters!');
+
             const musicOld = await musicRepository.fetchById(id);
-            musicOld.AlbumName = body.AlbumName ?? musicOld.AlbumName;
-            musicOld.MusicID = body.MusicID ?? musicOld.MusicID;
-            musicOld.MusicName = body.MusicName ?? musicOld.MusicName;
-            musicOld.MusicTitle = body.MusicTitle ?? musicOld.MusicTitle;
-            musicOld.MusicPoster = body.MusicPoster ?? musicOld.MusicPoster;
-            musicOld.MusicURL = body.MusicURL ?? musicOld.MusicURL;
-            musicOld.MusicDuration = body.MusicDuration ?? musicOld.MusicDuration;
-            musicOld.MusicLyrics = body.MusicLyrics ?? musicOld.MusicLyrics;
-            musicOld.MusicTags = body.MusicTags ?? musicOld.MusicTags;
-            musicOld.MusicArtists = body.MusicArtists ?? musicOld.MusicArtists;
-            musicOld.MusicReleaseTime = body.MusicReleaseTime ?? musicOld.MusicReleaseTime;
+            musicOld.AlbumName = body["album-name"] ?? musicOld.AlbumName;
+            musicOld.MusicName = body.name ?? musicOld.MusicName;
+            musicOld.MusicTitle = body.title ?? musicOld.MusicTitle;
+            musicOld.MusicPoster = body.poster ?? musicOld.MusicPoster;
+            musicOld.MusicURL = body.url ?? musicOld.MusicURL;
+            musicOld.MusicDuration = body.duration ?? musicOld.MusicDuration;
+            musicOld.MusicLyrics = body.lyrics ?? musicOld.MusicLyrics;
+            musicOld.MusicTags = body.tags ?? musicOld.MusicTags;
+            musicOld.MusicArtists = body.artists ?? musicOld.MusicArtists;
+            musicOld.MusicReleaseTime = body["release-time"] ?? musicOld.MusicReleaseTime;
 
             const music = await musicRepository.update(musicOld, req.UserID);
             if (!music)
@@ -142,7 +143,7 @@ class MusicController {
             else
                 sendResponse(res, statusCode.NO_CONTENT, { "Content-Type": "application/json" }, JSON.stringify(this.#print([music])));
         } catch (error) {
-            logger.error('updateMusic: ' + error);
+            logger.error(`${req.url}: ${error}`);
             throw error;
         }
     };
@@ -150,13 +151,16 @@ class MusicController {
     deleteMusic = async (req, res) => {
         try {
             const { id } = req.querystring;
+            if (!id)
+                return sendResponse(res, statusCode.BAD_REQUEST, { "Content-Type": "application/json" }, 'Invalid parameters!');
+
             const music = await musicRepository.delete(id, req.UserID);
             if (!music)
                 sendResponse(res, statusCode.GONE, { "Content-Type": "application/json" }, 'Could Not Delete!');
             else
                 sendResponse(res, statusCode.OK, { "Content-Type": "application/json" }, JSON.stringify(this.#print([music])));
         } catch (error) {
-            logger.error('deleteMusic: ' + error);
+            logger.error(`${req.url}: ${error}`);
             throw error;
         }
     };
